feat: Handler Exceptions

+++ src/main/java/com/example/meu_primeiro_springboot/service/ProductService.java
@@ -1,5 +1,6 @@
 package com.example.meu_primeiro_springboot.service;
 
+import com.example.meu_primeiro_springboot.exceptions.ResourceNotFoundException;
 import com.example.meu_primeiro_springboot.model.Product;
 import com.example.meu_primeiro_springboot.repository.ProductRepository;
 import org.springframework.stereotype.Service;
@@ -20,8 +21,9 @@ public class ProductService {
         return productRepository.findAll();
     }
 
-    public Optional<Product> getById(Long id) {
-        return productRepository.findById(id);
+    public Product getById(Long id) {
+        return productRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Produto com ID " +id+" não encontrado!"));
     }
 
     public Product saveProduct(Product product) {
@@ -29,6 +31,10 @@ public class ProductService {
     }
 
     public void deleteProduct(Long id){
+
+        if (!productRepository.existsById(id)){
+            throw new ResourceNotFoundException("Product with ID" + id + "Not Found!");
+        }
         productRepository.deleteById(id);
     }
 }
